% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM.R
\name{CPM}
\alias{CPM}
\title{CPM}
\usage{
CPM(matrix, log = FALSE, pseudocount = FALSE)
}
\arguments{
\item{matrix}{a matrix of counts, rows for features and columns for samples}

\item{log}{boolean whether to log2 transform the cpm}

\item{pseudocount}{boolean whether to add a pseudocount of 1 to the matrix}
}
\value{
a normalised count matrix
}
\description{
A function to generate cpm from expression matrix
}
\examples{
CPM(mtcars)

CPM(mtcars, log = TRUE)

CPM(mtcars, log = TRUE, pseudocount = TRUE)
}
\keyword{cpn}
\keyword{expression}
\keyword{normalisation}
