% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FancyUmap.R
\name{FancyUmap}
\alias{FancyUmap}
\title{umap_arrows}
\usage{
FancyUmap(
  seuratObj,
  reduction = "umap",
  group = "orig.ident",
  title = NULL,
  subtitle = NULL,
  split = NULL
)
}
\arguments{
\item{seuratObj}{a seurat object}

\item{reduction}{the reduction to use for the Dimplot (default: umap)}

\item{group}{the column in the metadata to colour the cells by (default: orig.ident)}

\item{title}{a title for the plot, if NULL a generic title will be added (default: NULL)}

\item{subtitle}{a subtitle for the plot, if NULL a generic subtitle will be added (default: NULL)}

\item{split}{a column in the metadata to split the umap into multiple umaps (default: NULL) N.B. this will break the arrows!}
}
\value{
a ggplot object
}
\description{
umap_arrows
}
\examples{

FancyUmap(seuratObj, reduction = "umap", group = "seurat_clusters", title = "test plot", subtitle = "a plot to test this function", split = NULL)

}
\keyword{UMAP}
\keyword{dimplot}
