% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FancyUmap.R
\name{FancyUmapPlot}
\alias{FancyUmapPlot}
\title{FancyUmapPlot}
\usage{
FancyUmapPlot(seuratObj, title = NULL, subtitle = NULL, ...)
}
\arguments{
\item{seuratObj}{a seurat object}

\item{title}{a title for the plot, if NULL a generic title will be added (default: NULL)}

\item{subtitle}{a subtitle for the plot, if NULL a generic subtitle will be added (default: NULL)}

\item{...}{Any other arguments that will be passed onto dimplot}
}
\value{
a ggplot object
}
\description{
FancyUmapPlot
}
\examples{

FancyUmapPlot(seuratObj)

FancyUmapPlot(seuratObj, title = "test plot", subtitle = "a plot to test this function")

}
\keyword{UMAP}
\keyword{dimplot}
