% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellHelpers.R
\name{PlotSeuratQCFromList}
\alias{PlotSeuratQCFromList}
\title{PlotSeuratQCFromList}
\usage{
PlotSeuratQCFromList(seuratList)
}
\arguments{
\item{seuratList}{a list of seurat objects}
}
\value{
a ggplot plot
}
\description{
A function to plot QC metrics from a list  of seurat objects
Useful if you're stuck thinking of a palette
}
\examples{
PlotSeuratQCFromList()

PlotSeuratQCFromList(seuList)
}
\keyword{scRNAseq}
