% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomColours.R
\name{randomColours}
\alias{randomColours}
\title{randomColours}
\usage{
randomColours(number = 10, height = 5)
}
\arguments{
\item{number}{number of colours to pick}

\item{height}{number of rows for plot}
}
\value{
a ggplot plot
}
\description{
A function to randomly select n colours and plot them.
Useful if you're stuck thinking of a palette
}
\examples{
randomColours()

randomColours(number = 57, height = 8)
}
\keyword{colour}
