% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM.R
\name{TPM}
\alias{TPM}
\title{TPM}
\usage{
TPM(matrix, length, log = TRUE, pseudocount = TRUE)
}
\arguments{
\item{matrix}{a matrix of counts, rows for features and columns for samples}

\item{length}{a vector of gene lengths in the same order as the matrix}

\item{log}{boolean whether to log2 transform the cpm}

\item{pseudocount}{boolean whether to add a pseudocount of 1 to the matrix}
}
\value{
a normalised count matrix
}
\description{
A function to generate tpm from expression matrix
}
\examples{
TPM(matrix, lengths)

}
\keyword{expression}
\keyword{normalisation}
\keyword{tpm}
