% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_arrows.R
\name{umap_arrows}
\alias{umap_arrows}
\title{umap_arrows}
\usage{
umap_arrows(
  plot,
  percentage = 0.2,
  label_x = "UMAP 1",
  label_y = "UMAP 2",
  x_pad = 0.2,
  y_pad = 0.2
)
}
\arguments{
\item{plot}{a ggplot object}

\item{percentage}{a value between 0 and 1 for how far on each axis the arrows should reach}

\item{label_x}{a string for the x arrow text}

\item{label_y}{a string for the y arrow text}

\item{x_pad}{padding for the x arrow label (percentage of arrow length)}

\item{y_pad}{padding for the y arrow label (percentage of arrow length)}
}
\value{
a ggplot object
}
\description{
umap_arrows
}
\examples{
p <- ggplot(mtcars, aes(x = mpg, y = wt)) +
    geom_point()
umap_arrows(p)
}
\keyword{UMAP}
\keyword{dimplot}
